old-tests/submission/Infraware/Forms/contents/Forms/select_type.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>select_type</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action=""
		  id="input_form">
		<p>
		 <label for="unittype">Select unit type:</label>
		 <select id="unittype" name="unittype">
		  <option value="1"> Miner </option>
		  <option value="2"> Puffer </option>
		  <option value="3" selected> Snipey </option>
		  <option value="4"> Max </option>
		  <option value="5"> Firebot </option>
		 </select>
		</p>	
	</form>

	<script>
		
		var select = document.getElementById("unittype");
		var type = select.type;
		select.multiple = true;

		if (typeof(select.type) == "string") {
			if (type == 'select-one' && select.type == 'select-multiple') {
				test(function() {
					assert_true(true, "type attribute is correct.");
				});
			} else {
				test(function() {
					assert_unreached("type attribute is not correct.");
				});
			}
		} else {
			test(function() {
				assert_unreached("type attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
